Snapshot resolution works now.
Its not really beautiful: I just wanted to add the snapshot repository
to the repository list of the artifact to be resolved, but this
wouldn't work. The maven meta data was correctly fetched from the
snapshot repository, but the actual artifact was then requested
from the releases repository which returned a 404 with the error message
that snapshots are forbidden there (of course). This can be seen
easily when switching to the debug log level. Well, it works now.