Artifact loading can now be done per component on request.
Loading all Maven Artifact at once might be nice for the reaction time of the application, but it takes very long when a lot of components are involved. Since most components won't be used with any individual pipeline, an on-demand loading of artifacts has been introduced to reduce startup times of applications.