Added 'forEditing' parameter to the load method of the JCoReUIMAPipeline.
When set to 'true', the serialized meta information is loaded. Since
this requires all respective classes - and their dependencies - to
reside on the classpath, it fails when using the pipeline runner that
tries to minimize dependencies outside of those the pipeline requires.
So the runner sets this parameter to 'false'.