Fixed a bug where non-AAE pipeline parts would not load properly.
When loading a pipeline, we want the descriptor files to be read with
custom changes made outside of the pipeline builder.
This didn't work with non-AAE components. I.e. when there was only
a single AE or a single consumer, the original file descriptor would
not be set but the serialized descriptor would be used.