Changing the JCoreGithubInformationService into a singleton.
The issue is that the service has a field for the local Maven
repository and the service itself is a field of MetaDescription.
When a pipeline is created on host A and loaded on host B, the
loading program may crash due to the fact that on B the local Maven
repository might differently be placed.