Fixing the issue with deprecated Maven SNAPSHOTS.
The point is that in the meta description previously we replaced
the MavenArtifact with the resolved artifact from a maven repository.
The issue with that is that snapshots may have a timestamp in their
version, exactly determining which snapshot version this is. We then
stored this exact version with the serialized pipeline data.
This means that when loading the pipeline, instead of requesting some
version 2.3.0-SNAPSHOT, for example, we specifically requested version
2.3.0-20180726.084412-37, even if we had updated this snapshot again
(to, say, version -38). Thus, we would always get the version that
had been used when the pipeline was created and saved instead of a
newer snapshot versions.