Made important enhancements in artifact version control.
First of all, when loading a pipeline, not found artifacts are now
tried to resolve to the newest available version. This can help
recover old pipelines built on snapshot versions.
Second, there is a new dialog in the configuration section that allows
to choose from all available versions for all component artifacts.