Reorder bug fix, configuration enhancements, code cleaning.
There was a bug when reordering multiple components of the same
component template, e.g. multiple Gazetteers. The internally used
'List.indexOf' operation could not distinguish between the
different components even if they had different configurations. This
was due to the implementation of the Description's "equals"
method. This method has been extended by almost each field of the
class which seems to have the desired effect.

Also, it is now possible to rename components and resources. The latter
is very important because UIMA does not handle multiple resources with
the same name.

Removed the JcoreGithubInformationService instance from the
MetaDescription which had been causing trouble there before and
was deprecated already.