Cleaning up code. Dependency and conflict resolution seems to work fine.
The conflict resolution apparently happens automatically when
resolving all dependencies together in a single request as we
did in the last commits. At least one case of a pipeline that had
conflicting libraries doesn't show the old behaviour any more. That means
that before the change there were multiple files for the same library
in different versions in the lib/ directory, now they aren't any more.