Revoked a change that rendered pipelines non-runnable.
In an attempt to remove unnecessary code, I removed the loading
of th CR from descriptor because I wanted to rely on the serialized
data. But this doesn't work when running the pipeline because we
then don't load the serialized data. This is because the serialization
classes are not strictly needed to run the pipeline and we want
to reduce dependencies of the pipeline runner as much as possible
to not interfere with the pipeline code.