Restructured the pipeline runner.
The goal is to have the new "CPEBootstrapRunner" calling the new
"CPERunner" via Runtime.exec(). The base issue is that the
pipeline runner's dependencies regularly interfere with the
executed pipeline. The newest example of this is the fact that some
dependency of the runner includes the Woodstox StAX parser. It is not
contained as a Maven dependency but the classes are just contained
in some dependency JAR, rendering a pipeline that requires another
StAX implementation unrunnable. That is why we are no going for
the bootstrap process where the original runner just retrieved the
classpath elements of the pipeline and than calls the CPE runner.