Fixed issues when the same component has multiple instances in the pipeline.
First, when the multiple component have the same name in a pipeline,
this pipeline cannot be loaded. In the method that sets the
XML descriptors to the descriptions, unique names are required.
This has been fixed by: Disallowing renaming components if the name
already exists, renaming components automatically upon addition into
the pipeline if necessary and refusing to save a pipeline with
duplicate component names. Some of these points are redundant to each
other but in this way, the user always gets immediate feedback.

Second, when setting a component version, all other instances of
the same component get the new version set, too.

When adding new components to the pipeline, the artifact version of the
new description is synchronized with already existing descriptions of
the same component, if available.