Trying to speed up usage of the pipeline builder CLI.
The most time takes the loading of libraries onto the classpath and
the retrieval of maven artifacts. We did this at each startup because for the creation
of aggregates, imports were resolved including those by name, i.e.
those located in JARs. However, now we do not load any JARs any more
and for some reason there are no errors. Perhaps they only occur
in specific cases or the latest refactoring of aggregate creation did
actually remove the import resolution.

Also, the dependencies are no longer stored every time the pipeline
is saved. A new field in the PipelineBuilderCLI keeps track of
component changes and version changes. Only if a change occurred,
the dependencies are stored into the lib/ directory. In case this
fails in certain circumstances, the saving dialog have been changed
to force the population of the lib directory.