The JCoRePipeline can now have a parent POM.
This parent POM will be used as the dependency tree root when resolving
the component dependencies. By specifying the dependencyManagement
section, one can fix library versions which would cause conflicts
otherwise (e.g. by using an older version of a library although the
newer one would be required but one component specifies the older and
another the newer version as a dependency).