Preventing the type imports of descriptors being resolved within the descriptors.
The issue is that when loading an existing pipeline and storing it again,
the UIMA descriptors suddenly do not reference their type systems
by imports but just include all the times right in the descriptor.
This was caused by retrieving the descriptors from the AAE descriptor
when loading. Somehow the descriptor there is completely resolved.
This is now circumvented by using the loaded descriptors directly when
setting them to the Description objects.