Trying to correct the count of processed documents in the CPE finishing report by supporting the multiplier types.
When there are multipliers, a single original CAS only carries URLs or other pointers to the actual documents to be read by the multiplier(s). Thus, while only a single CAS arrives at the `StatusCallbackListener`, it actually represents a number of documents. Now supporting the JCoReUris and RowBatch JCoRe multiplier types.