Support component AAEs as only components.
When there is only one AAE, this AAE was assumed to be the centralizing AAE created when storing pipelines. Thus, it was not checked if there is a component description for it. This is now fixed: When there is a component description whose name matches the AAE name, its descriptor is set to the description.