Update aether tools.
Because the old version did not use the local repository which could be very annoying. I did try to use the local repository two years ago, the code was still there in the aether tools. Back then I seemed to have issues with it. It worked well in a short test right now. Perhaps I was confused because of the error message "could not be found in repository LOCAL" when artifact resolution fails. However, it seems that remote repositories are still tried, the error message does just not mention them.